Moving the test router2 folder by renaming it as the previous used router. Renaming the old one router-legacy
