+++ app/router-legacy/search-router.js
